refactor: move helper functions
